+++ server/src/main/java/org/opensearch/cluster/node/DiscoveryNode.java
@@ -34,6 +34,7 @@ package org.opensearch.cluster.node;
 
 import org.opensearch.Version;
 import org.opensearch.common.UUIDs;
+import org.opensearch.common.annotation.PublicApi;
 import org.opensearch.common.settings.Setting;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.core.common.io.stream.StreamInput;
@@ -64,8 +65,9 @@ import static org.opensearch.node.NodeRoleSettings.NODE_ROLES_SETTING;
 /**
  * A discovery node represents a node that is part of the cluster.
  *
- * @opensearch.internal
+ * @opensearch.api
  */
+@PublicApi(since = "1.0.0")
 public class DiscoveryNode implements Writeable, ToXContentFragment {
 
     static final String COORDINATING_ONLY = "coordinating_only";